ENH fix circular import
