+++ pyccl/core.py
@@ -17,7 +17,6 @@ Some relevant data structures include:
 from . import ccllib as lib
 import numpy as np
 from warnings import warn
-from .pyutils import check
 
 # Configuration types
 transfer_function_types = {
@@ -635,3 +634,47 @@ class Cosmology(object):
 
         # Return status information
         return "status(%s): %s" % (status, msg)
+
+
+def check(status, cosmo=None):
+    """Check the status returned by a ccllib function.
+
+    Args:
+        status (int or :obj:`core.error_types`): Flag or error describing the
+                                                 success of a function.
+
+    """
+    # Check for normal status (no action required)
+    if status == 0:
+        return
+
+    # Get status message from Cosmology object, if there is one
+    if cosmo is not None:
+        msg = _cosmology_obj(cosmo).status_message
+    else:
+        msg = ""
+
+    # Check for known error status
+    if status in error_types.keys():
+        raise RuntimeError("Error %s: %s" % (error_types[status], msg))
+
+    # Check for unknown error
+    if status != 0:
+        raise RuntimeError("Error %d: %s" % (status, msg))
+
+
+def _cosmology_obj(cosmo):
+    """Returns a ccl_cosmology object, given an input object which may be
+    ccl_cosmology, the Cosmology wrapper class, or an invalid type.
+
+    Args:
+        cosmo (ccl_cosmology or Cosmology): The input cosmology which gets
+                                            converted to a ccl_cosmology.
+
+    """
+    if isinstance(cosmo, lib.cosmology):
+        return cosmo
+    elif isinstance(cosmo, Cosmology):
+        return cosmo.cosmo
+    else:
+        raise TypeError("Invalid Cosmology or ccl_cosmology object.")
