+++ pyccl/correlation.py
@@ -8,7 +8,7 @@ Choices of algorithms used to compute correlation functions:
 
 from . import ccllib as lib
 from . import constants as const
-from .pyutils import _cosmology_obj, check
+from .core import _cosmology_obj, check
 import numpy as np
 
 correlation_methods = {
