+++ pyccl/pyutils.py
@@ -2,34 +2,7 @@
 well as wrappers to automatically vectorize functions."""
 from . import ccllib as lib
 import numpy as np
-from . import core
-
-
-def check(status, cosmo=None):
-    """Check the status returned by a ccllib function.
-
-    Args:
-        status (int or :obj:`core.error_types`): Flag or error describing the
-                                                 success of a function.
-
-    """
-    # Check for normal status (no action required)
-    if status == 0:
-        return
-
-    # Get status message from Cosmology object, if there is one
-    if cosmo is not None:
-        msg = _cosmology_obj(cosmo).status_message
-    else:
-        msg = ""
-
-    # Check for known error status
-    if status in core.error_types.keys():
-        raise RuntimeError("Error %s: %s" % (core.error_types[status], msg))
-
-    # Check for unknown error
-    if status != 0:
-        raise RuntimeError("Error %d: %s" % (status, msg))
+from .core import check, _cosmology_obj
 
 
 def debug_mode(debug):
@@ -50,23 +23,6 @@ def debug_mode(debug):
         lib.set_debug_policy(lib.CCL_DEBUG_MODE_ON)
     else:
         lib.set_debug_policy(lib.CCL_DEBUG_MODE_OFF)
-
-
-def _cosmology_obj(cosmo):
-    """Returns a ccl_cosmology object, given an input object which may be
-    ccl_cosmology, the Cosmology wrapper class, or an invalid type.
-
-    Args:
-        cosmo (ccl_cosmology or Cosmology): The input cosmology which gets
-                                            converted to a ccl_cosmology.
-
-    """
-    if isinstance(cosmo, lib.cosmology):
-        return cosmo
-    elif isinstance(cosmo, core.Cosmology):
-        return cosmo.cosmo
-    else:
-        raise TypeError("Invalid Cosmology or ccl_cosmology object.")
 
 
 def _vectorize_fn_simple(fn, fn_vec, x, returns_status=True):
